Revert "Onboardingscreen"
