--- lib/constants/onboarding_data.dart
@@ -1,11 +1,0 @@
-List<Map<String, String>> onboardingData = [
-  {
-    "image_url": "healthCheck.png",
-    "text": "To be a responsible donor, you must get a check-up."
-  },
-  {
-    "image_url": "puzzle.png",
-    "text": "Your blood type should be compatible with the receiver’s type."
-  },
-  {"image_url": "bloodbag.png", "text": "Donate your blood and save a life."},
-];
