+++ lib/screens/hello.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: Container(),
+      ),
+    );
+  }
+}
